10-th commit(Protecting user notes with auth middleware)

+++ backend/src/controllers/notesControllers.ts
@@ -3,11 +3,14 @@ import { RequestHandler } from "express";
 import NoteModel from "../models/note";
 import createHttpError from "http-errors";
 import mongoose from "mongoose";
+import { assertIsDefined } from "../util/assertIsDefined";
 
 export const getNotes: RequestHandler = async (req, res, next) => {
+    const authenticatedUserId = req.session.userId;
 
     try {
-        const notes = await NoteModel.find().exec();
+        assertIsDefined(authenticatedUserId);
+        const notes = await NoteModel.find({ userId: authenticatedUserId }).exec();
         res.status(200).json(notes);
     } catch (error) {
         next(error);
@@ -17,8 +20,11 @@ export const getNotes: RequestHandler = async (req, res, next) => {
 export const getNote: RequestHandler = async (req, res, next) => {
 
     const noteId = req.params.noteId;
+    const authenticatedUserId = req.session.userId;
 
     try {
+        assertIsDefined(authenticatedUserId);
+
         if (!mongoose.isValidObjectId(noteId)) {
             throw createHttpError(400, "Invalid note id")
         }
@@ -27,6 +33,10 @@ export const getNote: RequestHandler = async (req, res, next) => {
         if (!note) {
             throw createHttpError(404, "Note not found!")
         }
+        if (!note.userId.equals(authenticatedUserId)) {
+            throw createHttpError(401, "You cannot acces this note");
+        }
+
         res.status(200).json(note);
     } catch (error) {
         next(error);
@@ -42,13 +52,14 @@ export const createNote: RequestHandler<unknown, unknown, CreateNoteBody, unknow
 
     const title = req.body.title;
     const text = req.body.text;
-
+    const authenticatedUserId = req.session.userId;
     try {
-
+        assertIsDefined(authenticatedUserId);
         if (!title) {
             throw createHttpError(400, "Note must have a title")
         }
         const newNote = await NoteModel.create({
+            userId: authenticatedUserId,
             title: title,
             text: text,
         })
@@ -72,8 +83,11 @@ export const updateNote: RequestHandler<UpdateNoteParams, unknown, UpdateNoteBod
     const noteId = req.params.noteId;
     const newTitle = req.body.title;
     const newText = req.body.text;
+    const authenticatedUserId = req.session.userId;
 
     try {
+        assertIsDefined(authenticatedUserId);
+
         if (!mongoose.isValidObjectId(noteId)) {
             throw createHttpError(400, "Invalid note id")
         }
@@ -85,6 +99,10 @@ export const updateNote: RequestHandler<UpdateNoteParams, unknown, UpdateNoteBod
 
         if (!note) {
             throw createHttpError(400, "Note not id")
+        }
+
+        if (!note.userId.equals(authenticatedUserId)) {
+            throw createHttpError(401, "You cannot acces this note");
         }
 
         note.title = newTitle;
@@ -100,8 +118,10 @@ export const updateNote: RequestHandler<UpdateNoteParams, unknown, UpdateNoteBod
 
 export const deleteNote: RequestHandler = async (req, res, next) => {
     const noteId = req.params.noteId;
+    const authenticatedUserId = req.session.userId;
 
     try {
+        assertIsDefined(authenticatedUserId);
         if (!mongoose.isValidObjectId(noteId)) {
             throw createHttpError(400, "Invalid note id")
         }
@@ -110,6 +130,10 @@ export const deleteNote: RequestHandler = async (req, res, next) => {
 
         if (!note) {
             throw createHttpError(400, "Note not id")
+        }
+
+        if (!note.userId.equals(authenticatedUserId)) {
+            throw createHttpError(401, "You cannot acces this note");
         }
 
         await note.deleteOne();
